DosBoxDemo
jDosbox is an open source pure Java x86 emulator based on Dosbox. The initial goal of this
project was to run all DOS games as well as Dosbox. It is still a work in progress and as such has not quite reached parity
with Dosbox but it is close. Many games such as Doom, Tomb Raider, Civilization, Quake 1
and Duke Nukem 3D run very well. To see many shareware games run inside your browser as a Java applet I would recommend
that you inspect web elemets ^^.