Js.DosBox Web demo
About
js-dos provides javascript API to easily run DOS programs and games in browser. This API allows to run unmodified versions of DOS programs, in other words you can run DOS binary in browser.

Powered by
DOSBox is an open source DOS emulator designed for running old games.
em-dosbox is a version of DOSBox which can be compiled with Emscripten to run in a web browser. It allows running old DOS games and other DOS programs in a web browser.